+++ make/404.php
@@ -9,13 +9,13 @@ get_header();
 	<article class="error-404 not-found">
 		<header class="entry-header">
 			<h1 class="entry-title">
-				<?php _e( 'Oops! That page can&rsquo;t be found.', 'make' ); ?>
+				<?php esc_html_e( 'Oops! That page can&rsquo;t be found.', 'make' ); ?>
 			</h1>
 		</header>
 
 		<div class="entry-content">
 			<p>
-				<?php _e( 'Maybe try searching this website:', 'make' ); ?>
+				<?php esc_html_e( 'Maybe try searching this website:', 'make' ); ?>
 			</p>
 			<?php get_search_form(); ?>
 		</div>