+++ make/comments.php
@@ -13,7 +13,8 @@ endif;
 	<h3 class="comments-title">
 		<?php
 		printf(
-			_nx( 'One comment', '%1$s comments', get_comments_number(), 'comments title', 'make' ),
+			// Translators: this string appears as the title of the Comments section of a post.
+			esc_html( _n( 'One comment', '%1$s comments', get_comments_number(), 'make' ) ),
 			number_format_i18n( get_comments_number() )
 		);
 		?>
@@ -21,7 +22,7 @@ endif;
 
 		<?php if ( get_comment_pages_count() > 1 ) : ?>
 		<nav id="comment-nav-above" class="comment-navigation" role="navigation">
-			<span class="screen-reader-text"><?php _e( 'Comment navigation', 'make' ); ?></span>
+			<span class="screen-reader-text"><?php esc_html_e( 'Comment navigation', 'make' ); ?></span>
 			<?php paginate_comments_links(); ?>
 		</nav>
 		<?php endif; ?>
@@ -37,7 +38,7 @@ endif;
 
 		<?php if ( get_comment_pages_count() > 1 ) : ?>
 		<nav id="comment-nav-below" class="comment-navigation" role="navigation">
-			<span class="screen-reader-text"><?php _e( 'Comment navigation', 'make' ); ?></span>
+			<span class="screen-reader-text"><?php esc_html_e( 'Comment navigation', 'make' ); ?></span>
 			<?php paginate_comments_links(); ?>
 		</nav>
 		<?php endif; ?>
@@ -46,7 +47,7 @@ endif;
 
 	<?php if ( ! comments_open() && '0' != get_comments_number() && post_type_supports( get_post_type(), 'comments' ) ) : ?>
 	<p class="no-comments">
-		<?php _e( 'Comments are closed.', 'make' ); ?>
+		<?php esc_html_e( 'Comments are closed.', 'make' ); ?>
 	</p>
 	<?php endif; ?>