+++ make/partials/entry-date.php
@@ -10,7 +10,8 @@ $date_option = ttfmake_sanitize_choice( get_theme_mod( $date_key, ttfmake_get_de
 $date_string = get_the_date();
 if ( 'relative' === $date_option ) :
 	$date_string = sprintf(
-		_x( '%s ago', 'time period', 'make' ),
+		// Translators: this string indicates an amount of passed time. e.g. 5 minutes ago
+		__( '%s ago', 'make' ),
 		human_time_diff( get_the_time( 'U' ), current_time( 'timestamp' ) )
 	);
 endif;